Refactor prebuilt configurations
Split prebuilt configurations into separate files for better organization and maintainability.
[skip gpt_engineer]
